refactor: file structure
- MyFetch, MyPromise 독립적으로 사용하기 위해 외부 의존성 제거
