+++ MyFetch/constants/errorMessages.js
@@ -1,0 +1,4 @@
+export default {
+  DATA_TYPE: `Data type must be Object.`,
+  NO_URL: `URL is neccessary.`
+}
