+++ MyFetch/constants/typeStrings.js
@@ -1,0 +1,8 @@
+export default {
+  FUNCTION: `function`,
+  NUMBER: `number`,
+  OBJECT: `object`,
+  BOOLEAN: `boolean`,
+  STRING: `string`,
+  UNDEFINED: `undefined`
+}
