+++ MyFetch/index.js
@@ -1,20 +1,20 @@
 import MyPromise from "../MyPromise/index.js";
-import errMsg from "../constants/ErrorMessages.js";
-import type from "../constants/TypesString.js";
+import errMsg from "./constants/errorMessages.js";
+import type from "./constants/typeStrings.js";
 
 const defaultOptions = {
   method: "GET"
 }
 
 function MyFetch(url, options) {
-  if(!url) new Error(errMsg.fetch.NO_URL);
+  if(!url) new Error(errMsg.NO_URL);
   const mergedOption = {...defaultOptions, ...options};
 
   return new MyPromise((resolve, reject) => {
     const req = new XMLHttpRequest();
     req.addEventListener("load", () => {
       let data = JSON.parse(req.responseText);
-      if(typeof data !== type.OBJECT) reject(errMsg.fetch.DATA_TYPE);
+      if(typeof data !== type.OBJECT) reject(errMsg.DATA_TYPE);
       else resolve(data);
     });
     req.open(mergedOption.method, url);
