--- constants/ErrorMessages.js
@@ -1,10 +1,0 @@
-export default {
-  promise : {
-    RESOLVER_TYPE: `Promise resolver must be a function.`
-  },
-
-  fetch : {
-    DATA_TYPE: `Data type must be Object.`,
-    NO_URL: `URL is neccessary.`
-  }
-}
