--- constants/TypesString.js
@@ -1,8 +1,0 @@
-export default {
-  FUNCTION: "function",
-  NUMBER: "number",
-  OBJECT: "object",
-  BOOLEAN: "boolean",
-  STRING: "string",
-  UNDEFINED: "undefined"
-}
